feat: add enhanced link render hook
Applies to Markdown links. With

[render_hooks]
[render_hooks.link]
staticFallback = true

(the default) in params.toml (or not present), behaviour falls back to
current behaviour of taking supplied link verbatim, after attempting to
resolve as below. With `false` attempts to resolve as below and emits an
error or warning if the link cannot be found.

Links are first resolved as content page, then page resources, then
global, and if no such resource exists. Default falls back to using the
link verbatim, but this can be disabled (which helps avoid broken
internal links).

Email links are handled as before; they are obfuscated.

Code originally by Veriphor and modified by various contributors

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>